Configured constants moved to their own file
