+++ src/main/java/frc/robot/subsystems/DriveData.java
@@ -2,7 +2,7 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj.motorcontrol.Spark;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.Constants;
+import frc.robot.Configured;
 
 public class DriveData {
 
@@ -12,15 +12,15 @@ public class DriveData {
     public Spark backRightDrive;
 
     public DriveData() {
-        this.frontLeftDrive = new Spark(Constants.DRIVE_FRONT_LEFT_PORT);
-        this.frontRightDrive = new Spark(Constants.DRIVE_FRONT_RIGHT_PORT);
-        this.backLeftDrive = new Spark(Constants.DRIVE_BACK_LEFT_PORT);
-        this.backRightDrive = new Spark(Constants.DRIVE_BACK_RIGHT_PORT);
+        this.frontLeftDrive = new Spark(Configured.DRIVE_FRONT_LEFT_PORT);
+        this.frontRightDrive = new Spark(Configured.DRIVE_FRONT_RIGHT_PORT);
+        this.backLeftDrive = new Spark(Configured.DRIVE_BACK_LEFT_PORT);
+        this.backRightDrive = new Spark(Configured.DRIVE_BACK_RIGHT_PORT);
 
-        this.frontLeftDrive.setInverted(Constants.DRIVE_FRONT_LEFT_FLIPPED);
-        this.frontRightDrive.setInverted(Constants.DRIVE_FRONT_RIGHT_FLIPPED);
-        this.backLeftDrive.setInverted(Constants.DRIVE_BACK_LEFT_FLIPPED);
-        this.backRightDrive.setInverted(Constants.DRIVE_BACK_RIGHT_FLIPPED);
+        this.frontLeftDrive.setInverted(Configured.DRIVE_FRONT_LEFT_FLIPPED);
+        this.frontRightDrive.setInverted(Configured.DRIVE_FRONT_RIGHT_FLIPPED);
+        this.backLeftDrive.setInverted(Configured.DRIVE_BACK_LEFT_FLIPPED);
+        this.backRightDrive.setInverted(Configured.DRIVE_BACK_RIGHT_FLIPPED);
 
 
     }
