+++ src/main/java/frc/robot/subsystems/InputData.java
@@ -4,7 +4,7 @@ package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
-import frc.robot.Constants;
+import frc.robot.Configured;
 import frc.robot.functions.TelemetryUtil;
 
 public final class InputData {
@@ -15,8 +15,8 @@ public final class InputData {
 
 
     public InputData() {
-        controller = new XboxController(Constants.CONTROLLER_PORT);
-        joystick = new Joystick(Constants.JOYSTICK_PORT);
+        controller = new XboxController(Configured.CONTROLLER_PORT);
+        joystick = new Joystick(Configured.JOYSTICK_PORT);
     }
